+++ src/lib/components/InputField.svelte
@@ -27,7 +27,7 @@
 	on:change={() => {
 		dispatch('change', value);
 	}}
-	class="ml-auto flex-grow rounded-lg border border-slate-500 border-opacity-0 bg-gray-900 px-3
+	class="flex-grow rounded-lg border border-slate-500 border-opacity-0 bg-gray-900 px-3
 		 py-1 text-slate-300 placeholder-slate-400 invalid:ring-2 invalid:ring-red-500
 		  valid:hover:border-opacity-100 focus:border-opacity-0 focus:outline-none focus:ring-2 valid:focus:ring-green-400
 		  disabled:text-slate-400 text-{size} placeholder:text-{size} {className}"